# docs/swagger.yaml
basePath: /
definitions:
  genproto.Categories:
    properties:
      Categories:
        items:
          $ref: '#/definitions/genproto.Category'
        type: array
    type: object
  genproto.Category:
    properties:
      CategoryId:
        type: string
      Name:
        type: string
    type: object
  genproto.Comment:
    properties:
      Body:
        type: string
      CommentId:
        type: string
      PostId:
        type: string
      UserId:
        type: string
    type: object
  genproto.CommentModel:
    properties:
      Body:
        type: string
      PostId:
        type: string
      UserId:
        type: string
    type: object
...
  genproto.Post:
    properties:
      Body:
        type: string
      CategoryId:
        type: string
      PostId:
        type: string
      Title:
        type: string
      UserId:
        type: string
    type: object
...
  genproto.Posts:
    properties:
      Posts:
        items:
          $ref: '#/definitions/genproto.Post'
        type: array
    type: object
...
info:
  contact: {}
  title: Online Forum API Documentation
paths:
...
  /category/{id}:
    put:
      consumes:
      - application/json
      description: This api Updates Category by Name
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Category'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Updates Category
      tags:
      - Category
  /category/{name}:
    delete:
      consumes:
      - application/json
      description: This api Deletes Category by Name
      parameters:
      - description: Category Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: category deleted successfully!
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Deletes Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: This api Gets Category by Name
      parameters:
      - description: Category Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Category'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Category
      tags:
      - Category
  /category/all:
    get:
      consumes:
      - application/json
      description: This api Gets all Categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Categories'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets All Categories
      tags:
      - Category
  /category/create:
    post:
      consumes:
      - application/json
      description: This api create Category
      parameters:
      - description: Category Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Category'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Creates Category
      tags:
      - Category
  /category/posts/{id}:
    get:
      consumes:
      - application/json
      description: This api Gets Posts by Category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Posts'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Posts by Category
      tags:
      - Category
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: This api Deletes Commnet
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Deletes Comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: This api Gets Commnet
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: This api Updates Commnet
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.CommentModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Updates Comment
      tags:
      - Comment
  /comment/all:
    get:
      consumes:
      - application/json
      description: This api Gets All Commnet
      parameters:
      - description: Post ID
        in: query
        name: post_id
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Post Body
        in: query
        name: body
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets All Comment
      tags:
      - Comment
  /comment/create:
    post:
      consumes:
      - application/json
      description: This api create Commnet
      parameters:
      - description: Comment Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.CommentModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Creates Comment
      tags:
      - Comment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"